# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Error is a short, human-readable description of what went\nwrong. The server never returns raw stack traces or DB\nerror strings here; the goal is to be informative without\nleaking internals.",
                        "example": "Invalid request payload",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "JWT bearer token issued by /api/auth/login",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/vexgo-org/vexgo/issues"
        },
        "description": "Self-hosted blog CMS HTTP API.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "termsOfService": "https://github.com/vexgo-org/vexgo",
        "title": "VexGo API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/sso/{provider}/callback": {
            "get": {
                "description": "Handles the redirect-back from the provider. On\nsuccess it returns an HTML page that writes the\nresult to localStorage and closes the popup; on\nfailure the same shape is used with an \"error\"\nfield. The frontend listens for the storage event\nunder the key `sso_callback_result` to pick up\nthe data.",
                "parameters": [
                    {
                        "description": "provider slug",
                        "in": "path",
                        "name": "provider",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "state nonce from the original /login redirect",
                        "in": "query",
                        "name": "state",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "authorization code from the provider",
                        "in": "query",
                        "name": "code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "HTML — success popup closer"
                    },
                    "400": {
                        "description": "HTML — error popup closer"
                    }
                },
                "summary": "OAuth2 callback endpoint",
                "tags": [
                    "sso"
                ]
            }
        },
        "/sso/{provider}/login": {
            "get": {
                "description": "Redirects the browser to the OAuth2 provider's\nauthorization endpoint. The `method` query\nparameter controls the post-callback behaviour:\n` sso_get_token` (default) issues a JWT for full\nlogin; `get_sso_id` only returns the\nprovider-side user id (used to bind SSO to an\nexisting account).",
                "parameters": [
                    {
                        "description": "provider slug (github, google, ...)",
                        "in": "path",
                        "name": "provider",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "flow variant",
                        "in": "query",
                        "name": "method",
                        "schema": {
                            "enum": [
                                "sso_get_token",
                                "get_sso_id"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "redirect to the provider's authorization URL"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "unknown provider or method"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Start an OAuth2 SSO flow",
                "tags": [
                    "sso"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}